requirements.txt
streamlit==1.28.1
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
httpx==0.27.2
